+++ pro/mcp/core/src/main/kotlin/org/http4k/mcp/protocol/messages/McpCancelled.kt
@@ -1,16 +1,17 @@
 package org.http4k.mcp.protocol.messages
 
+import org.http4k.mcp.model.McpMessageId
 import org.http4k.mcp.model.Meta
 import org.http4k.mcp.protocol.McpRpcMethod
 import se.ansman.kotshi.JsonSerializable
 
-object Cancelled : McpRpc {
+object McpCancelled : McpRpc {
     override val Method = McpRpcMethod.of("notifications/cancelled")
 
     @JsonSerializable
     data class Notification(
-        val requestId: String,
-        val reason: String?,
+        val requestId: McpMessageId,
+        val reason: String? = null,
         override val _meta: Meta = Meta.default,
     ) : ClientMessage.Notification, ServerMessage.Notification, HasMeta
 }