+++ pro/mcp/core/src/main/kotlin/org/http4k/mcp/util/McpJson.kt
@@ -1,5 +1,3 @@
 package org.http4k.mcp.util
 
-import org.http4k.format.MoshiNode
-
 object McpJson : ConfigurableMcpJson()